+++ components/Note/Note.js
@@ -1,6 +1,6 @@
 import { Box, Group, Stack, Text } from "@mantine/core";
 import { Kind } from "nostr-tools";
-import React, { useMemo, useState } from "react";
+import React, { useEffect, useMemo, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { CommentIcon, ShakaIcon, ZapIcon } from "../../icons/StemstrIcon";
 import { useProfile } from "../../ndk/hooks/useProfile";
@@ -15,8 +15,12 @@ import useStyles from "./Note.styles";
 import { useRouter } from "next/router";
 import { openSheet } from "store/Sheets";
 import { useNote } from "ndk/hooks/useNote";
+import { useNDK } from "ndk/NDKProvider";
+import { formatETag, parseEventTags } from "ndk/utils";
+import { NDKEvent } from "@nostr-dev-kit/ndk";
 
 const Note = (props) => {
+  const { ndk } = useNDK();
   const { event, type } = props;
   const note = useNote({ event });
   const { classes } = useStyles();
@@ -33,10 +37,54 @@ const Note = (props) => {
       note.event.tags?.find((tag) => tag[0] === "download_url") || null;
     return downloadUrlTag ? downloadUrlTag[1] : null;
   }, [note]);
+  const [likedByCurrentUser, setLikedByCurrentUser] = useState(false);
 
   const handleClickComment = (e) => {
     dispatch(openSheet({ sheetKey: "postSheet", replyingTo: note.event }));
   };
+
+  const handleClickLike = (e) => {
+    let created_at = Math.floor(Date.now() / 1000);
+    const { root, mentions, reply } = parseEventTags(note.event);
+
+    let eTags = [];
+    if (root) eTags.push(root);
+    eTags.push(...mentions);
+    eTags.push(formatETag(["e", note.event.id], "reply"));
+    if (reply) eTags.push(formatETag(reply, "mention"));
+
+    let pTags = note.event.tags.filter((t) => t[0] === "p");
+    pTags.push(["p", note.event.pubkey]);
+
+    let tags = [...eTags, ...pTags];
+
+    const reactionEvent = new NDKEvent(ndk);
+    reactionEvent.kind = Kind.Reaction;
+    reactionEvent.created_at = created_at;
+    reactionEvent.tags = tags;
+    reactionEvent.content = "❤️";
+    reactionEvent
+      .publish()
+      .then(() => {
+        setLikedByCurrentUser(true);
+        controls.start({
+          scale: [1, 1.25, 1],
+          transition: {
+            duration: 0.5,
+            ease: "easeInOut",
+          },
+        });
+      })
+      .catch((err) => {});
+  };
+
+  useEffect(() => {
+    if (!likedByCurrentUser) {
+      if (note.reactions.find((ev) => ev.pubkey === auth.pk)) {
+        setLikedByCurrentUser(true);
+      }
+    }
+  }, [note.reactions.length, auth.pk]);
 
   const handleClick = () => {
     router.push(`/thread/${note.event.id}`);
@@ -75,7 +123,11 @@ const Note = (props) => {
           <Group position="apart">
             <NoteActionComment note={note} onClick={handleClickComment} />
             {/* <RepostButton note={note} /> */}
-            <NoteActionLike note={note} />
+            <NoteActionLike
+              note={note}
+              onClick={handleClickLike}
+              likedByCurrentUser={likedByCurrentUser}
+            />
             {/* <NoteAction>
             <ZapIcon width={18} height={18} /> 4
           </NoteAction> */}
